writting stuff

+++ Core/CTFAK.Core/CCN/Chunks/Frame/Events/GroupPointer.cs
@@ -20,7 +20,7 @@ namespace CTFAK.MMFParser.EXE.Loaders.Events.Parameters
         public override void Write(ByteWriter Writer)
         {
             Writer.WriteInt32(Pointer);
-            Writer.WriteInt32(Id);
+            Writer.WriteInt16(Id);
         }
     }
 }
